split seq/set/map types into separate files
