+++ Source/DafnyCore/AST/Types/CollectionType.cs
@@ -109,112 +109,3 @@ public abstract class CollectionType : NonProxyType {
   /// </summary>
   public abstract CollectionBoundedPool GetBoundedPool(Expression source);
 }
-
-public class SetType : CollectionType {
-  public bool Finite { get; }
-
-  [SyntaxConstructor]
-  public SetType(IOrigin? origin, bool finite, List<Type?> typeArgs) : base(origin, typeArgs) {
-    Finite = finite;
-  }
-
-  public SetType(bool finite, Type? arg) : base(arg) {
-    Finite = finite;
-  }
-
-  public override string CollectionTypeName => Finite ? "set" : "iset";
-
-  [System.Diagnostics.Contracts.Pure]
-  public override bool Equals(Type that, bool keepConstraints = false) {
-    var t = that.NormalizeExpand(keepConstraints) as SetType;
-    return t != null && Finite == t.Finite && Arg!.Equals(t.Arg, keepConstraints);
-  }
-
-  public override Type Subst(IDictionary<TypeParameter, Type> subst) {
-    var arg = Arg?.Subst(subst);
-    if (arg is InferredTypeProxy) {
-      ((InferredTypeProxy)arg).KeepConstraints = true;
-    }
-    return arg == Arg ? this : new SetType(Finite, arg);
-  }
-
-  public override Type ReplaceTypeArguments(List<Type> arguments) {
-    return new SetType(Finite, arguments[0]);
-  }
-
-  // Sets always support equality, because there is a check that the set element type always does.
-  public override bool SupportsEquality => true;
-
-  public override BinaryExpr.ResolvedOpcode ResolvedOpcodeForIn => BinaryExpr.ResolvedOpcode.InSet;
-  public override CollectionBoundedPool GetBoundedPool(Expression source) {
-    return new SetBoundedPool(source, Arg, Arg, Finite);
-  }
-}
-
-public class MultiSetType : CollectionType {
-  [SyntaxConstructor]
-  public MultiSetType(IOrigin? origin, List<Type?> typeArgs) : base(origin, typeArgs) { }
-
-  public MultiSetType(Type? arg) : base(arg) { }
-
-  public override string CollectionTypeName => "multiset";
-
-  public override bool Equals(Type that, bool keepConstraints = false) {
-    var t = that.NormalizeExpand(keepConstraints) as MultiSetType;
-    return t != null && Arg!.Equals(t.Arg, keepConstraints);
-  }
-
-  public override Type Subst(IDictionary<TypeParameter, Type> subst) {
-    var arg = Arg?.Subst(subst);
-    if (arg is InferredTypeProxy) {
-      ((InferredTypeProxy)arg).KeepConstraints = true;
-    }
-    return arg == Arg ? this : new MultiSetType(arg);
-  }
-
-  public override Type ReplaceTypeArguments(List<Type> arguments) {
-    return new MultiSetType(arguments[0]);
-  }
-
-  // Multisets always support equality, because there is a check that the set element type always does.
-  public override bool SupportsEquality => true;
-
-  public override BinaryExpr.ResolvedOpcode ResolvedOpcodeForIn => BinaryExpr.ResolvedOpcode.InMultiSet;
-  public override CollectionBoundedPool GetBoundedPool(Expression source) {
-    return new MultiSetBoundedPool(source, Arg, Arg);
-  }
-}
-
-public class SeqType : CollectionType {
-  [SyntaxConstructor]
-  public SeqType(IOrigin? origin, List<Type?> typeArgs) : base(origin, typeArgs) { }
-
-  public SeqType(Type? arg) : base(arg) { }
-
-  public override string CollectionTypeName => "seq";
-
-  public override bool Equals(Type that, bool keepConstraints = false) {
-    var t = that.NormalizeExpand(keepConstraints) as SeqType;
-    return t != null && Arg!.Equals(t.Arg, keepConstraints);
-  }
-
-  public override Type Subst(IDictionary<TypeParameter, Type> subst) {
-    var arg = Arg?.Subst(subst);
-    if (arg is InferredTypeProxy) {
-      ((InferredTypeProxy)arg).KeepConstraints = true;
-    }
-    return arg == Arg ? this : new SeqType(arg);
-  }
-
-  public override Type ReplaceTypeArguments(List<Type> arguments) {
-    return new SeqType(arguments[0]);
-  }
-
-  // The sequence type supports equality if its element type does
-  public override bool SupportsEquality => Arg!.SupportsEquality;
-
-  public override BinaryExpr.ResolvedOpcode ResolvedOpcodeForIn => BinaryExpr.ResolvedOpcode.InSeq;
-  public override CollectionBoundedPool GetBoundedPool(Expression source) {
-    return new SeqBoundedPool(source, Arg, Arg);
-  }
-}
